# app/swagger.yaml
openapi: 3.0.0
info:
  title: Employee Clocking System API
  description: API documentation for the employee clocking system.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /auth/login:
    post:
      summary: User Login
      description: Authenticate an employee.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Unauthorized
...
  /employees/{id}/attendance:
    get:
      summary: Get Employee Attendance
      description: Retrieve attendance records of an employee.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attendance data retrieved
        '404':
          description: Employee not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []